docker version/apimsa/requirements.txt
fastapi==0.88.0
pydantic==1.10.2
numpy==1.24.0
pandas==1.5.2
scipy==1.9.3
matplotlib==3.6.2
arabic-reshaper==2.1.4
python-bidi==0.4.2
uvicorn==0.19.0